[FEAT] 빌딩 특이사항 추가 삭제 api 추가

+++ src/main/java/com/propertyservice/property_service/repository/property/BuildingRemarkRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface BuildingRemarkRepository extends JpaRepository<BuildingRemark, Integer> {
+public interface BuildingRemarkRepository extends JpaRepository<BuildingRemark, Long> {
     List<BuildingRemark> findAllByBuilding(Building building);
 }
